\name{RcmdrPlugin.orloca-internal}
\title{RcmdrPlugin.orloca: A GUI for orloca-package (internal functions)}
\alias{activeDataSetLocaP}
\alias{Rcmdr.contour.zsum}
\alias{Rcmdr.help.orloca}
\alias{Rcmdr.help.RcmdrPlugin.orloca}
\alias{Rcmdr.new.loca.p}
\alias{Rcmdr.orloca.norm}
\alias{Rcmdr.persp.zsum}
\alias{Rcmdr.plot.contour.loca.p}
\alias{Rcmdr.plot.loca.p}
\alias{Rcmdr.plot.zsumn}
\alias{Rcmdr.rloca.p}
\alias{Rcmdr.summary.loca.p}
\alias{Rcmdr.zsum}
\alias{Rcmdr.zsummin}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.orloca packages. 
  For user information ask for help about \code{\link[RcmdrPlugin.orloca]{RcmdrPlugin.orloca}}.}
\keyword{internal}
