\name{RcmdrPlugin.orloca-internal}
\title{RcmdrPlugin.orloca: A GUI for orloca-package (internal functions)}
\alias{activeDataSetLocaP}
\alias{Rcmdr.contour.distsum}
\alias{Rcmdr.help.orloca}
\alias{Rcmdr.help.orloca.vignettes}
\alias{Rcmdr.help.RcmdrPlugin.orloca}
\alias{Rcmdr.help.RcmdrPlugin.orloca.vignettes}
\alias{Rcmdr.new.loca.p}
\alias{Rcmdr.orloca.norm}
\alias{Rcmdr.persp.distsum}
\alias{Rcmdr.plot.contour.loca.p}
\alias{Rcmdr.plot.loca.p}
\alias{Rcmdr.plot.distsumn}
\alias{Rcmdr.rloca.p}
\alias{Rcmdr.summary.loca.p}
\alias{Rcmdr.distsum}
\alias{Rcmdr.distsummin}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for RcmdrPlugin.orloca packages.}
\keyword{internal}
